fix: cast paged model to FP16 not BF16 to match stock HF baseline
Paged engine's load_hybrid_paged_qwen cast the model to torch.bfloat16 after
to_empty(), but stock HF's trust_remote_code path casts to torch.float16.
Over 28 layers this produced systematic rounding drift (~20% L2 inflation
at the final layer, final-layer cosine 0.94 instead of 0.99), and at the
logits projection this shifted "Paris" from top-1 to rank 4 on the prompt
"The capital of France is".

Probe evidence (mega_autopilot_day17/phase1_int8/drift_fp16.json):
  Before fix (BF16): paged top-5 = [' ____', ' ______', ' __', ' Paris', '____']
  After fix (FP16): paged top-5 = [' Paris', ' __', ' ______', ' ____', '____']
  Stock HF top-5:   [' Paris', ' __', ' a', ' ____', ' ______']

Final-layer cosine: BF16 → 0.939, FP16 → 0.989.

Regression test added: tests/test_v32_paged_paris.py (3 tests covering
top-1, top-5, and non-degenerate outputs).

Completes OUTLIER-DAY17-MEGA-AUTOPILOT-72H Phase 1 primary gate.
Secondary int8 shared-expert quality is no longer the bottleneck —
FP16 fixes the observed symptom cleanly.

+++ outlier_engine/paging.py
@@ -942,7 +942,12 @@ def load_hybrid_paged_qwen(
             # down_proj` safetensors keys.
 
     model = model.to_empty(device=torch.device(device))
-    model = model.to(dtype=torch.bfloat16)
+    # V3.2 safetensors store FP32 weights. Stock HF's trust_remote_code path
+    # casts to FP16 (matches the modeling file's computations). Previously the
+    # paged loader cast to BF16 which accumulates ~20% L2 drift over 28 layers
+    # and shifts logit ordering — e.g. "Paris" drops from top-1 to rank 4.
+    # Matching stock HF with FP16 closes that drift.
+    model = model.to(dtype=torch.float16)
 
     # `to_empty()` materialises every parameter on `device` but does NOT copy
     # the original data — nn.Parameters that were constructed under
